+++ clikt/src/jvmTest/kotlin/com/github/ajalt/clikt/parameters/PromptOptionsTest.kt
@@ -1,15 +1,20 @@
 package com.github.ajalt.clikt.parameters
 
+import com.github.ajalt.clikt.core.BadParameterValue
 import com.github.ajalt.clikt.core.subcommands
 import com.github.ajalt.clikt.parameters.options.check
 import com.github.ajalt.clikt.parameters.options.nullableFlag
 import com.github.ajalt.clikt.parameters.options.option
 import com.github.ajalt.clikt.parameters.options.prompt
+import com.github.ajalt.clikt.parameters.types.int
 import com.github.ajalt.clikt.testing.TestCommand
+import com.github.ajalt.clikt.testing.formattedMessage
 import com.github.ajalt.clikt.testing.parse
 import com.github.ajalt.clikt.testing.test
+import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.shouldBe
 import org.junit.Rule
+import org.junit.contrib.java.lang.system.SystemErrRule
 import org.junit.contrib.java.lang.system.SystemOutRule
 import org.junit.contrib.java.lang.system.TextFromStandardInputStream
 import kotlin.test.Test
@@ -64,6 +69,19 @@ class PromptOptionsTest {
         C().parse("")
         stdout.logWithNormalizedLineSeparator shouldBe "Foo: Bar: "
     }
+
+    @Test
+    fun `prompt option after error`() {
+        class C : TestCommand(false) {
+            val foo by option().int()
+            val bar by option().prompt()
+        }
+        shouldThrow<BadParameterValue> {
+            C().parse("--foo=x")
+        }.formattedMessage shouldBe "invalid value for --foo: x is not a valid integer"
+        stdout.logWithNormalizedLineSeparator shouldBe ""
+    }
+
 
     @Test
     fun `prompt flag`() {